Feat : [SHK-141] 피드 최근 날짜 순 조회 기능 구현
- (중요) 조회 성능을 높이기 위해 좋아요를 필드로 가지도록 엔티티, DDL과 관련 코드 수정
- 다른 도메인에서 수정된 스키마(Order, Transaction)까지 통합 반영
- QueryDSL 저장, 수정, 삭제 외 조회에 QueryDSL 적용

+++ src/main/java/com/prgrms/kream/domain/style/dto/response/GetFeedFacadeResponse.java
@@ -7,7 +7,7 @@ public record GetFeedFacadeResponse(
 		Long id,
 		Long authorId,
 		String content,
-		Long likes,
+		Integer likes,
 		List<String> images,
 		LocalDateTime createdAt,
 		LocalDateTime updatedAt
